app/requirements.txt
pandas==0.25.3
striprtf==0.0.5
requests
spacy
matplotlib
wordcloud
gensim==3.8.1
pprint
camelot-py[cv]
